requirements-dev.txt
# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-watch>=4.2.0

# Linting & Formatting
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0

# Project Dependencies
setuptools>=65.0.0
wheel>=0.37.0
python-dotenv
openai
PyPDF2>=3.0.0 